Convert Storage Module to TS
